IntrinsicsGenX.td: Allow single bit mask on wrregion
Wrregion's final argument is a mask governing which elements of the
subvector will be written into the overall vector. This will be used as
part of the implementation of SIMD control flow.

Previously, this mask had to be a vector of i1 with the same width as
the subvector. This was rather unwieldy if the subvector was large, and
mask functionality would not be required in such a case anyway because
the mask would never be bigger than 32.

With this change, the mask argument is allowed to be a single i1, which
is considered to the the value of all elements of the mask. Since we
have not tried to implement SIMD control flow yet, all existing example
code can just use "i1 1" here.

Only the declaration of llvm.genx.wrregion is changed here. No
implementation code needs to be changed because nothing takes any notice
of the mask yet.

Change-Id: Ia555973d55d5bc912e3dd082e2ce4c1a96911926
Reviewed-on: http://mlpgit002.isw.intel.com/63
Reviewed-by: Stuttard, David <david.stuttard@intel.com>
Tested-by: Stuttard, David <david.stuttard@intel.com>