IntrinsicsGenX.td: Slight change to llvm.genx.wrregion
I have changed the comment defining llvm.genx.wrregion to say that the
mask argument is a vector of i1 _at least_ as wide as the subvector,
rather than _exactly_ as wide. We may need this flexibility when we
sort out how perdicated ops get split up by legalization. It will also
allow GenXLowering to get rid of a 0 index rdpredregion entirely if its
only use is in a select being lowered to a wrregion.

Change-Id: Ib227cae4e0c5215e12d1a406319780dc6ace9bdb
Reviewed-on: http://mlpgit002.isw.intel.com/104
Reviewed-by: Stuttard, David <david.stuttard@intel.com>
Tested-by: Stuttard, David <david.stuttard@intel.com>