GenXCategoryConversion new pass
A category conversion is an instruction that converts a value from one
category to another, e.g. general to address, or general to surface
index.

Previously, a category conversion was represented by a bitcast. Code in
GenXVisaFunctionWriter could see the categories of the allocated
registers and generate the appropriate vISA instruction. These category
conversion bitcasts were added in GenXLiveness, except for conversion to
address register, which was added in GenXAddressConversion run after
GenXCoalescing so it could see when multiple uses of an address derived
from the same index also used the same base register so could be
commoned up.

The problem with adding category conversions that late is that it stops
a category conversion baling anything in. In sepia I saw an
example where a conversion to address register (a vISA addr_add
instruction) should bale in a rdregion. This example shows something
that would be very common: for a vector of bytes (so no scaling of the
index), the addr_add is likely to want to bale in the rdelement that
comes from lowering the trunc to i16 of the index.

This change moves the addition of category conversion instructions into
a new pass GenXCategoryConversion that is run before the second (final)
run of GenXBaling. Also, a category conversion is represented by a new
llvm.genx.convert intrinsic, except for conversion to address, which
gets its own llvm.genx.convert.addr intrinsic (representing an addr_add
vISA instruction).

Adding address conversion earlier means that we can no longer see
whether multiple address uses of the same index value can be commoned
up, so GenXCategoryConversion assumes that each use needs its own
llvm.genx.convert.addr call.

A subsequent change will reinstate commoning up of address conversions
where possible in a later pass.

Change-Id: Icfe51b026e166cc225bbcdf5d71ae29b8af2e3ce
Reviewed-on: http://mlpgit002.isw.intel.com/70
Reviewed-by: Stuttard, David <david.stuttard@intel.com>
Tested-by: Stuttard, David <david.stuttard@intel.com>