GenX: New llvm.genx.convert.pred intrinsic
Declared and implemented llvm.genx.convert.pred intrinsic, which
converts an i32 into a v32i1 predicate variable. It generates a vISA
setp instruction.

Doing this included implementing the ability for an intrinsic to have a
predicate destination operand.

The new intrinsic will be used in GenXConstantMaterialization to
materialize a predicate constant. I was going to use a bitcast, but I
discovered that CSE removes the bitcast and puts the constant back into
the instruction that uses it.

Change-Id: Ifd77fd73d30171b3cd4e2f7310b66f3866ad8a55
Reviewed-on: http://mlpgit002.isw.intel.com/117
Reviewed-by: Stuttard, David <david.stuttard@intel.com>
Tested-by: Stuttard, David <david.stuttard@intel.com>