GenX: More address conversion changes
GenXCategoryConversion adds address conversion instructions
(llvm.genx.convert.addr) where necessary.

The pass started with a value, and looked at its uses to see if any of
them was as an index in a region access. Because we want to bale a
constant add of an index in to the region using it, when it found a
constant add as a use, it had to look through it to try and find region
index uses.

All this became a bit complicated where a constant add was used both as
a region index and in another add which itself was used as a region
index. The pass went wrong and inserted two address conversions, one
before each add.

This commit rehashes how GenXCategoryConversion handles an add with an
index use -- it turns it into an llvm.genx.add.addr intrinsic, so when
the pass gets to the value that the add uses, it is obvious that the add
is an address use without having to look through it.

The commit includes some other changes needed for this to work:

1. GenXCategoryConversion now processes instructions in reverse order of
dominance, so it will see the add and convert it to llvm.genx.add.addr
before it sees what the add uses and insert an llvm.genx.convert.addr.

2. GenXBaling now supports llvm.genx.convert.addr, and bales it in as an
address add if possible. It still needs to support normal add/sub, as
GenXBaling is run twice, and the first time is before
GenXCategoryConversion.

3. GenXVisaFunctionWriter now supports an unbaled llvm.genx.add.addr
(generating an add_addr vISA instruction) for any case where the
llvm.genx.add.addr fails to be baled in to a region access.

Change-Id: I60d1c2aaa86794d1fa249832851d449fef369acc
Reviewed-on: http://mlpgit002.isw.intel.com/134
Reviewed-by: Stuttard, David <david.stuttard@intel.com>
Tested-by: Stuttard, David <david.stuttard@intel.com>