GenX: Use rdregion/wrregion for scalar region
Following Wei C's suggestion, this commit removes rdelement and
wrelement intrinsics, and instead uses rdregion and wrregion to
represent read/write of a scalar in a vector after lowering from
extractelement/insertelement.

Patchset 2: In the Region constructor from a rdregion/wrregion
instruction, had a bit more of a tidy up and fixed a bug where it was
reading the wrong operand from wrregion.

Change-Id: I7f03266bfe6a96d18c95682303062355beec2944
Reviewed-on: http://mlpgit002.isw.intel.com/156
Reviewed-by: Stuttard, David <david.stuttard@intel.com>
Tested-by: Stuttard, David <david.stuttard@intel.com>