GenX: Removed constant materialization from GenXCoalescing
A future change is needed to allow GenXCoalescing, when it adds a copy
after failing to coalesce, to cope with needing to split that copy
because it is too big or not a power of 2 size.

This commit is intended to simplify the work in that future change by
moving any constant materialization out of GenXCoalescing, and
implementing the insertion of a constant copy for an intrinsic operand
that is not allowed to be immediate. All such constant copies are now
inserted in GenXLowering, using a new llvm.genx.constant intrinsic to
protect the constant from being propagated by CSE. Any such constant
move is then legalized in terms of its size by GenXLegalization, and
legalized in terms of what constant it is moving by
GenXConstantMaterialization.

This leaves GenXCoalescing needing to deal with inserting only
non-constant moves.

Change-Id: If56ba1c7f6db89b53810582dcf4c4fcab8968a05
Reviewed-on: http://mlpgit002.isw.intel.com/195
Reviewed-by: Liu, Zhenying <zhenying.liu@intel.com>
Tested-by: Liu, Zhenying <zhenying.liu@intel.com>
Reviewed-by: Stuttard, David <david.stuttard@intel.com>
Tested-by: Stuttard, David <david.stuttard@intel.com>