GenX: unary operators and source modifier baling
Declared and implemented llvm.genx.absf and llvm.genx.absi intrinsics.

In baling, detect abs, neg (0-x) and not (x^-1) as source modifiers that
can be baled in.

In vISA writing, added support for source modifiers, whether baled in to
something else or not.

This code is pretty much untested other than a little play with abs and
negate.

Change-Id: I9f4a860bcb9fd177f333c9c9af86a76e49d88a14
Reviewed-on: http://mlpgit002.isw.intel.com/339
Tested-by: sys_Jenkins LD <sys_jenkins.ld@intel.com>
Reviewed-by: Stuttard, David <david.stuttard@intel.com>