GenX: Changes to shared function intrinsics
media.ld and oword.ld had a dummy arg with the same type as the return
type, because I thought that would help overload resolution in Clang. It
turns out we do not need them, so I have now removed them.

media.ld and media.st had args for block width and height. The block
height is unnecessary as it can be inferred from the width and the type
of the data being read/written, so I have removed it.

oword.ld and oword.st had an arg for the number of owords. It too can be
inferred from the type of the data being read/written, so I have removed
it.

This change is interdependent with the corresponding changes in Clang
and in the .ll tests. As such it will not verify automatically.

Change-Id: I90fc0d0f778975bede25c1944379a1e0effd2d6b
Reviewed-on: http://mlpgit002.isw.intel.com/343
Reviewed-by: Stuttard, David <david.stuttard@intel.com>
Tested-by: Renouf, Tim <tim.renouf@intel.com>