IntrinsicsGenX.td: wrregion now allows v32i1 mask
I have changed the spec of wrregion so the mask argument is allowed to
be a v32i1. If the number of elements in the region is more than
32, then the mask is replicated as required.

I have not implemented this in the GenX backend yet.

Also added to the spec that the element type must be a power of two
number of bytes up to and including 8 bytes, so i1 is not allowed.

Change-Id: I38cb4b9c3db90cec22c2255a81b801c6b2b1e01b
CI-Depends: I38cb4b9c3db90cec22c2255a81b801c6b2b1e01b
Reviewed-on: http://ctmgit002.isw.intel.com/694
Tested-by: sys_Jenkins LD <sys_jenkins.ld@intel.com>
Reviewed-by: Pan, Wei <wei.pan@intel.com>