GenX: constant predicate changes
A constant predicate is materialized using an intrinsic (rather than a
bitcast, which would allow the constant to be propagated back again in
the CSE pass). This commit changes the name of the intrinsic from
convert_pred to constantpred, to better reflect its use, and also
ensures that the resulting value has the right type rather than always
v32i1.

Change-Id: I6796178c0f7327b8c628afab014319d93bf55d40
Reviewed-on: http://ctmgit002.isw.intel.com/741
Reviewed-by: Stuttard, David <david.stuttard@intel.com>
Tested-by: sys_Jenkins LD <sys_jenkins.ld@intel.com>