GenX: Better lowering of not
Bitwise not is represented in LLVM IR by xor with all ones. This commit
adds GenX backend internal intrinsics llvm.genx.noti and llvm.genx.notp.
GenXLowering lowers to them, so it is easier to recognize noti as a
source operand modifier and to generate notp to a vISA not instruction.

Change-Id: If4656898a87e66d403b47041521687f625b63c79
Reviewed-on: http://ctmgit002.isw.intel.com/742
Reviewed-by: Stuttard, David <david.stuttard@intel.com>
Tested-by: sys_Jenkins LD <sys_jenkins.ld@intel.com>