GenXCoalescing: fix use of genx_convert instead of bitcast
When GenXCoalescing needs to insert a copy, it uses bitcast for a value
of general category, or the genx_convert intrinsic if a conversion is
required.

This commit fixes it so it uses genx_convert for any non-general
category, even if a conversion is not required. Also changed the spec of
genx_convert and the part of GenXVisaFuncWriter where its vISA is
written to allow for this.

This fixes an invalid vISA problem in STDE once the fix to propagate
categories through subroutine args is in.

Change-Id: Ifa7c8d406254b679cdad94845bea115c9c4c655a
CI-Depends: Ifa7c8d406254b679cdad94845bea115c9c4c655a
Reviewed-on: http://ctmgit002.isw.intel.com/16020
Tested-by: sys_Jenkins LD <sys_jenkins.ld@intel.com>
Reviewed-by: Stuttard, David <david.stuttard@intel.com>