GenX: tidied handling of predicate constants
This commits makes handling of predicate constants more like other
constants.

Insertion of code to load a predicate constant is now done in
GenXLowering rather than GenXConstantMaterialization.

The llvm.genx.constantpred intrinsic now takes a vector of i1 constant
as its operand, rather than the constant converted to an i32. The
conversion to i32 now happens in GenXVisaFuncWriter when writing the
setp instruction.

GenXLegalization now copes with llvm.genx.constantpred and splits it if
necessary. I have added code to cope with the case that a predicated
wrregion does not need to be split, but its predicate is wider than
necessary (which is allowed), and the predicate's web needs splitting.

Patchset 3: rebase

Change-Id: Ic69f279dbb018592848d356bb8e3436ddb56a628
CI-Depends: Ic69f279dbb018592848d356bb8e3436ddb56a628
Reviewed-on: http://ctmgit002.isw.intel.com/16090
Tested-by: sys_Jenkins LD <sys_jenkins.ld@intel.com>
Reviewed-by: Stuttard, David <david.stuttard@intel.com>