GenXIntrinsics: line and pln changes
I had defined the line and pln intrinsics as needing a "FIXED4" operand
0, meaning that it is always a 4 vector.

I recently discovered that, although that's how it works in the cm_line
and cm_pln intrinsics, in vISA you actually have a normal width vector
operand and it only uses elements 0, 1 and 3.

Therefore I have changed the definition of the llvm.genx.line and
llvm.genx.pln intrinsics to take a normal width vector (i.e. same as the
result width) for src0, with the following qualification:

    The coefficients for the pln operation are arg0[4n+0], arg0[4n+1]
    and arg0[4n+3]. Which n you get for each 8 wide part of the
    operation is undefined, because of the effect of legalization in the
    GenX backend.

    Therefore, to get the correct effect, arg0 should be the result of a
    rdregion<0;4,1> (i.e. replicate rows of 4 elements) whose input is a
    4 wide vector, so the same arg0[0], arg0[1] and arg0[3] are
    guaranteed to be used.

(And analogously for line.)

This has allowed me to remove the FIXED4 operand restriction in
GenXIntrinsics, and any code that uses it in GenXLegalization and
GenXBaling.

Also, this commit fixes a typo in the GenXIntrinsics table entries for
line and pln.

Change-Id: I00e0fb080bd8aa0acbebaebb125cf296a07dd367
CI-Depends: I00e0fb080bd8aa0acbebaebb125cf296a07dd367
Reviewed-on: http://ctmgit002.isw.intel.com/16445
Reviewed-by: Stuttard, David <david.stuttard@intel.com>
Tested-by: sys_Jenkins LD <sys_jenkins.ld@intel.com>