Don't try to predicate 1-vector stores in SIMD CF lowering.
SIMD CF lowering attempts to treat 1-vector stores specially
and avoids predicating them, considering them to be really
scalars. However, it was only detecting these properly if
the destination of the store was the input operand of a
wrregion. If it was not, it ended up trying to convert the
store into a select instruction with a condition vector the
same size as the SIMD width - which asserted in debug builds
because that was never going to be a 1-vector.

Fix this by simply treating 1-vector stores as always being
unpredicable under all circumstances.

Change-Id: If4ce6c4cf895b09f80d6c68b82df8c15cdac45d4
CI-Depends: I551609e15c75cdc127dab49dfee7849394640a9d
CI-Depends: If4ce6c4cf895b09f80d6c68b82df8c15cdac45d4
Reviewed-on: http://ctmgit002.isw.intel.com/16676
Tested-by: sys_Jenkins LD <sys_jenkins.ld@intel.com>
Reviewed-by: Renouf, Tim <tim.renouf@intel.com>