CMKernelArgOffset: early pass to determine offsets of kernel args
Previously the offsets of kernel args was done in GenXVisaFuncWriter.
The calculation requires knowledge of which implicit args are kernel and
which are thread, which is specific to the CM runtime and does not
really belong in the GenX backend.

This commit moves the calculation of kernel arg offsets into a new
CMKernelArgOffset pass run very early on. The offsets are now kept in
the kernel metadata for GenXVisaFuncWriter to use.

This introduces the possibility of a kernel arg getting an offset that
is not as aligned as required by some use of the arg, or something it is
coalesced with. So there is new code in GenXCoalescing to insert a copy
for a kernel arg if its offset is not aligned enough, after all
coalescing has occurred.

Change-Id: I039ab09d837787505e3102fb20264e9fd7fdcfea
CI-Depends: I039ab09d837787505e3102fb20264e9fd7fdcfea
CI-Depends: I1a2c5a9a68a89ff6979e0a39c8df367842acbabe
CI-Depends: I5dd4e7e9c9627310af9d8366ac6d78e53201db9f
CI-Depends: I31ef67143dc8a2eb30ee66e4f50ceb8e6e98fd59
CI-Depends: I6be5ddd7a58b6e110d95931ea36d482023235892
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/36020
Tested-by: cmllvm <cmllvm@intel.com>
Reviewed-by: Stuttard, David <david.stuttard@intel.com>